move Resolution from guilib to windowing
