+++ xbmc/cores/VideoPlayer/VideoRenderers/BaseRenderer.h
@@ -24,7 +24,7 @@
 #include <vector>
 
 #include "RenderInfo.h"
-#include "guilib/Resolution.h"
+#include "windowing/Resolution.h"
 #include "utils/Geometry.h"
 #include "VideoShaders/ShaderFormats.h"
 #include "cores/IPlayer.h"
