+++ xbmc/windowing/CMakeLists.txt
@@ -1,9 +1,11 @@
 set(SOURCES GraphicContext.cpp
             OSScreenSaver.cpp
+            Resolution.cpp
             WinSystem.cpp)
 
 set(HEADERS GraphicContext.h
             OSScreenSaver.h
+            Resolution.h
             WinEvents.h
             WinSystem.h
             XBMC_events.h
