+++ xbmc/windowing/Resolution.h
