+++ xbmc/windowing/rpi/RPIUtils.h
@@ -21,7 +21,7 @@
 #include "platform/linux/RBP.h"
 #include "EGL/egl.h"
 #include <bcm_host.h>
-#include "guilib/Resolution.h"
+#include "windowing/Resolution.h"
 
 class DllBcmHost;
 class CRPIUtils
